backend/requirements.txt
# backend
fastapi
fastapi_mail
uvicorn
# db
sqlalchemy
python-dotenv
psycopg2
alembic # migrations
